@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "avatara_brain_games"
-version = "0.1.1"
+version = "0.2.0"
 description = ""
 authors = ["Andrey <larin.projects@gmail.com>"]
 packages = [
@@ -8,7 +8,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.5"
+python = "^3.6"
 prompt = "^0.4.1"
 
 [tool.poetry.dev-dependencies]
@@ -17,8 +17,8 @@ pytest = "^5.4.3"
 
 [tool.poetry.scripts]
 brain-games = "brain_games.scripts.brain_games:main"
+brain-even = "brain_games.scripts.brain_even:main"
 
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
-
